refactor : section 컬렉션 일급 컬렉션으로 개선 - Line 에서 Sections를 사용하면 되므로 단위 테스트는 Line에서 하게 하였고,   Sections 사용은 Line을 이용해서만 하면 되기 때문에 Sections의 메서드들은 default 제약을 함.
