+++ src/main/java/nextstep/subway/applicaion/dto/LineResponse.java
@@ -36,7 +36,8 @@ public class LineResponse {
     }
 
     private static List<StationResponse> createStationResponses(Line line) {
-        if (line.getSections().isEmpty()) {
+//        if (line.getSections().isEmpty()) {
+        if (line.isEmptySections()) {
             return Collections.emptyList();
         }
